Soft blockage point, many questions.
